+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Planned
+
+- Deprecation of SolCyc
+
 ## [1.2.0-dev]
 
 ### Added
@@ -15,11 +19,11 @@
 - Pre-hooks uses now ruff instead of flake8 and black. This is reflected also
 in the dev-environment
 - Jupyter files for the documentation reflect current changes
-- Test data updated (15.01.24)
 - Dropped support for python 3.8 and 3.9
 - Logging format is more human-friendly
 - Dropped support of mypy. Black is replaced with ruff
 - New pyproject.toml
+- Test data updated (19.02.24)
 
 ### Removed